pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = ["ddcdatabases/**/*"]

[tool.hatch.build.targets.wheel]
packages = ["ddcdatabases"]

[project]
name = "ddcDatabases"
version = "4.0.0"
description = "Simplified database ORM connections with support for multiple database engines"
urls.Repository = "https://github.com/ddc/ddcDatabases"
urls.Homepage = "https://pypi.org/project/ddcDatabases"
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Daniel Costa", email = "daniel@ddcsoftwares.com"},
]
maintainers = [
    {name = "Daniel Costa"},
]
keywords = [
    "python", "python3", "python-3",
    "ORM", "ORM-connection", "sqlalchemy",
    "database", "databases", "ddcDatabase", "ddcDatabases",
    "mssql", "mssql-database",
    "mysql", "mysql-database",
    "oracle", "oracle-database",
    "mariadb", "mariadb-database",
    "mongodb", "mongodb-database",
    "sqlite", "sqlite3", "sqlite3-database",
    "postgres", "postgres-database", "postgresql", "postgresql-database", "pgsql", "pgsql-database",
]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Natural Language :: English",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic-settings>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.48",
]

[project.optional-dependencies]
mongodb = ["motor>=3.7.1"]
oracle = ["oracledb>=3.4.2"]
mssql = ["pyodbc>=5.3.0", "aioodbc>=0.5.0"]
mysql = ["mysqlclient>=2.2.8", "aiomysql>=0.3.2"]
postgres = ["psycopg[binary]>=3.3.3", "asyncpg>=0.31.0"]
pgsql = ["ddcdatabases[postgres]"]
mariadb = ["ddcdatabases[mysql]"]

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "poethepoet>=0.42.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
    "testcontainers[postgres,mysql,mssql,mongodb,oracle]>=4.14.1",
]

[tool.poe.tasks]
linter.shell = "uv run ruff check --fix . && uv run ruff format ."
profile = "uv run python -m cProfile -o cprofile_unit.prof -m pytest tests/unit"
profile-integration = "uv run python -m cProfile -o cprofile_integration.prof -m pytest tests/integration"
test.sequence = [{ shell = "uv run coverage run -m pytest tests/unit" }, { shell = "uv run coverage report" }, { shell = "uv run coverage xml" }]
test-integration = "uv run pytest tests/integration"
tests.sequence = ["linter", "test", "test-integration"]
updatedev.sequence = ["linter", {shell = "uv lock --upgrade && uv sync --all-extras --group dev"}]
build.sequence = ["updatedev", "tests", {shell = "uv build --wheel"}]

[tool.pytest.ini_options]
addopts = "-v --import-mode=importlib --junitxml=junit.xml"
junit_family = "legacy"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore:Field name \"schema\" in .* shadows an attribute in parent:UserWarning",
]
[tool.coverage.run]
omit = [
    "tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "S", "SLF"]
ignore = ["E501", "E402", "UP046", "UP047"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "S105", "S106", "S311", "SLF001", "F841"]

[tool.ruff.lint.isort]
known-first-party = ["ddcdatabases"]
no-lines-before = ["future", "standard-library", "third-party", "first-party", "local-folder"]
no-sections = true
